Move data builders into their own module for greater utility
